Only return active users
Part of the purpose of the idp-pw-api-personnel-* adapters is to prove
that a given user exists and is active in that personnel system. Our
various ID Store (aka Personnel system) adapters do not return inactive
users, and the PersonnelUser class returned by this
idp-pw-api-personnel-* adapter has no `active` field. Both of those facts
suggest we simply do not return inactive users here.